test: fix rust tests 20/20, regenerate readdir snapshots, fix CI paths
- Rust-тести приведено до 20/20 зелених (prro, offline, timeout, sync, e2e)
- Оновлено снапшоти редирів (categories/products/suppliers_default)
- Виправлено CI-пайплайн (.github/workflows/ci.yml)
- Додано scripts/regenerate_readdir_snapshots.py
- .gitignore: scripts/schema.sql (db dump)

+++ frontend/src-tauri/crates/torgashka-api/src/pos.rs
@@ -1504,7 +1504,7 @@ fn parse_cash_amount(v: &Value) -> Result<BigDecimal, PosErr> {
             &raw,
         )
     })?;
-    if amount <= BigDecimal::from(0) {
+    if amount <= 0 {
         return Err(v422(
             "greater_than",
             &["body", "amount"],
